Add fritzbox_thermostat_state metric
This adds a new metric fritzbox_thermostat_state representing the
thermostat state: 1/0 (on/off), -1 if unknown or error.
As the termostat state is privided via special values of the thermstats
goal temperature, the goal temperature can not be read if the termostate
is in off state (e.g. the metric will not show up/will not be updated in
that case).
Also fix a couple of error messages in case optional values could not be
read from thermostats.

Fixes #10